refactor: Folder / File restructure.
