+++ test/lib/resolved.js
@@ -1,4 +1,4 @@
-import { created as $created } from '../../src/util/symbols';
+import { created as $created } from 'src/util/symbols';
 
 export default function resolved (elem) {
   return elem[$created];
