+++ test/unit/lifecycle/attributes.spec.js
